+++ changelog.rst
@@ -34,6 +34,7 @@ Last change: 08-APR-2026 CEST
 - SPINNA allows user-defined threshold for the binary mask
 - Dialogs with scroll areas show no margins (e.g., Display settings dialog in Render)
 - Added help buttons to some dialogs/menu bars across the modules that open the corresponding readthedocs pages (the documentation will be further improved in the future)
+- "What's this?" help button removed from all dialogs (Windows)
 
 *Bug fixes:*
 ++++++++++++